paramgroup refactoring 1
