+++ src/WDCMethods.js
@@ -1,4 +1,4 @@
-import { get, multiGet } from "./utils.js";
+import { get, multiGet, combineJSONList } from "./utils.js";
 import { locationMode } from "./enums.js";
 import { notify } from "./notifications.js";
 import { parse, toSeconds } from "iso8601-duration";
@@ -557,24 +557,6 @@ const getSchema = schemaCallback => {
       })
       .catch(err => notify(err));
   }
-};
-
-/*
-This function amalgmates any number of input data JSON into single JSON, Modifies the first JSON in the list to become the result, and returns a reference to this JSON. Global
-metadata on all JSON except for the first JSON (with the exception of query URL) is assumed to be the same as the first and is not preserved, as this function is only intended to be used with batch queries specifically for the 
-purpose of bypassing the limit of 100 parameters on the instantaneous values services. This function must be called with a list of data JSON of length 2 or greater. 
-*/
-const combineJSONList = JSONList => {
-  let result = JSONList[0];
-  result.value.queryInfo.multi = true;
-  result.value.queryInfo.queryURL = [result.value.queryInfo.queryURL];
-  JSONList.slice(1).forEach(element => {
-    element.value.timeSeries.forEach(series => {
-      JSONList[0].value.timeSeries.push(series);
-      result.value.queryInfo.queryURL.push(element.value.queryInfo.queryURL);
-    });
-  });
-  return result;
 };
 
 export {
