+++ tests/unit/inputValidation.spec.js
@@ -7,7 +7,6 @@ import {
   validateSiteInputs,
   validateHydroCodeInputs,
   validateCountyInputs,
-  validateParamInputs,
   validateSiteTypeInputs,
   validateAgencyInputs,
   validateNatAquiferInput,
@@ -398,17 +397,6 @@ done in CountySelect.vue
     expect(validateCountyInputs(countyCode, wrapper.vm)).toBe(true);
   });
 
-  test("validateParamInputs successfully accepts valid params", () => {
-    const store = new Vuex.Store({
-      state: {},
-      modules: {},
-      getters: {},
-      actions: {}
-    });
-    const wrapper = shallowMount(Main, { store, localVue });
-    let paramCode = ["test"];
-    expect(validateParamInputs(paramCode, wrapper.vm)).toBe(true);
-  });
   test("validateSiteTypeInputs rejects an invalid site type", () => {
     const store = new Vuex.Store({
       state: {},
